Added some more property tests, added argument parsing from the cli, updated readme.

+++ background/headers/edge.hpp
@@ -5,11 +5,11 @@
 class Edge {
     public:
         // indices of v1 and v2
-        Edge(std::size_t v1, std::size_t v2, float length2, std::size_t identifier);
+        Edge(std::size_t v1, std::size_t v2, double length2, std::size_t identifier);
         std::size_t identifier;
         std::size_t v1Index;
         std::size_t v2Index;
-        float length2;
+        double length2;
         bool traversed = false;
         std::string toString() const noexcept;
         bool operator<(const Edge& other) const;
